Allow coalescence of precoloured nodes.
The code generator now loads function arguments like this:

    x = r1

In the common case, x should probably be allocated to r1, but there may
be cases where it can be useful to keep the move instruction.  To handle
this case I had to extend the coalescing code to handle precoloured nodes.

The change is straightforward.  If one of the nodes is precoloured we keep
it and make the other node point to this one.  If both nodes are precoloured
we don't coalesce.  (If they have different colours, we cannot coalesce; if
they have the same colour it wouldn't change anything for the neighbours.)