Add TAILCALL invariant for zero-copy function call.
Ensure that tailcall arguments are located in registers r0, r1, ...,
r(N-1).  This means that we don't need to copy in the interpreter, but
potentially in the bytecode.  Most of the time, the register allocator
seems to do a good job of removing moves, but there are a few issues
when swapping two registers.  For example:

    subtract dict x y = (Num.-) dict y x

The bytecode looks as follows:

    (alloc_frame 4)
    r3 = r2
    r2 = r1
    r4 = &GHC.Num.-_0Bn
    r1 = r3
    return r4(r0, r1, r2),

If we had a SWAP instruction and could detect it in the register
allocator.  This could be written shorter as:

    (alloc_frame 4)
    swap r1 r2
    r3 = &GHC.Num.-_0Bn
    return r3(r0, r1, r2),

These cases are probably rare, though, so we currently don't do this
optimisation.