Create HeapInfo to analyse structure of allocated data.
In order to do NEW-elimination we need to find out whether the allocation
is only needed when leaving the trace and can be omitted if otherwise.  That is,
similar to Snapshots, we only perform the allocation when we exit the trace.

However, we cannot do this if the purpose of the loop is to allocate data, so we
need to keep track of the contents of each allocated object.   The HeapInfo
structure is a first step to efficiently access the needed information.