Fewer live-outs for EVAL and CALL.
Both EVAL and CALL use MOV_RES to store the result of evaluation
in a result register.  The register for holding the result is therefore
not live immediately after the EVAL/CALL and can be marked as dead.
(This can help the JIT do more optimisations, because it needs to restore
fewer results.)