Add type information to 'BcVar' and use it to create GC pointer masks.
The bytecode compiler now keeps track of the type of each variable
(ignoring type variables and applications).  A few notes:

  - Polymorphic types are treated as pointers.
  - We do not currently remove zero-width variables.  E.g., 'State#'.

Pointer masks are not stored inline with the bytecode.  All pointer
and liveness masks are stored in the bitmap region at the end of the
bytecode region and instructions with masks store a relative pointer
to their entry of the bitmap region.