Tag static thunks as CAFs and update them differently.
Updated CAFs contain pointers into the live heap and must be treated
as roots.  We don't want to trace all static objects at each GC so we
use a write barrier.  Whenever, we update a static thunk we add the
updated CAF into a static_obj list.

Currently, we check for CAFs on each update.  This causes an extra
memory access, that would otherwise not be needed.  A future
improvement would instead push a different update continuation when we
start evaluating the thunk.  At that point we need to look at the info
table anyway.