Added initial implementation for entering traces
This commit contains a first shot at entering and exiting
traces. The code generator was updated to output a jump to
an exit stub. With this commit we can do a full roundtrip
from generating the (placeholder) machine code for a trace,
executing the machine code, going through an exit stub,
generating the exit state, and then returning to the interpreter.