Added codegen for LOOP instruction
We now correctly generate the backwards branch for a loop.

Correct codgen for resolving phi node mismatches.

Correct codgen for fixing phi nodes spilled outside loop
Phi nodes that are spilled only outside a loop need to have
IR_RENAMEs emitted for them.  We need to do this because if
the LHS was spilled outside of the loop the snapshot
restoration will attempt to load it from the spill slot.
This spill slot will be out of date as the value changes in
the loop.  We emit a rename to ensure that snapshots will
restore the value from the register inside the loop.