Change sps_scale to just scale the spill slot
The previous version was also subtracting SPS_FIRST so that
we could use all the spill slots. The problem is that the
spill area is zero indexed (i.e. spills[0] is the first
spill), but SPS_FIRST cannot be zero because a zero means
no spill slot is allocated. As a result we simply don't use
the first spill slot. This makes it easy to index into
spills[ir->s] without having to worry about re-adjusting the
index.