Use cost model for spilling regsiter in allocator
The cost to spill is a combination of the reference number
plus an extra bit for those values that flow to phi nodes.
Due to the organization of the IR, constants will have a
smaller cost than values defined outside the loop, which
will have a smaller cost than values defined inside the
loop. The extra phi-factor tries to ensure that values in
phi nodes will get a register.