Fixed error in FLOAD codegen
When the FLOAD was loading from a FREF whose first operand
was a constant, the code generated was incorrect. We now
correctly handle this case.

To make this work we moved all code for generating the
correct FREF inside the FLOAD. The `fuseload` function is
now unused so we delete it in this commit.

The Bench.SumFromTo3 test now passes successfully.