Fixed bug in register allocator
We were not properly handling the case where a value is live
on entry to the loop but dies in the loop and its register
gets reused for another value. The register would have the
wrong value after the first iteration of the loop.

The fix is to insert a load of the invariant at the top of
the loop when its register is modified in the body of the
loop.

The Bench.Primes test is now working.