Use the usage type of identifiers for calculating type info.
We got a compiler crash when trying to compile the expression:

  app :: (a -> b) -> a -> b
  test = (app eqInt 3) (4 :: Int)

The problem is that, per its polymorphic type, `app` only takes two
parameters.  However, when instantiating `b` with a function type
it may take more, as in the above case.

The fix is to look at the instantiated type of `app` when using its
type information.  The fix is currently nicely localised in the
`viewGhcApp` function.  Let's see if there are some hidden gotchas
I haven't encountered, yet.