Start splitting heap allocations into heap check + writes.
The new HEAPCHK instruction performs a single heap check for all
potential allocations during the next iteration.  Heap objects don't
bump the heap pointer, but instead simply initialize the memory
reserved by the previous heap check.

The heap check optimistically assumes that the trace will execute
fully.  If a trace exit is taken, this may waste some memory.  It
shouldn't be too hard to decrement the heap pointer in such cases,
but we defer that to future work.