Calculate trace framesize.
The framesize of a trace fragment is the maximum snapshot size
occurring within the trace.  While recording we may encounter deeper
stacks, but we can ignore those if no snapshot is generated at those
depths.

Note that the register allocator may create spill slots on the frame,
so code generation may increase the frame size.