Add self-verification mode and a few smaller changes.
The self-verification mode can only be used with the IR
interpreter.  It executes the IR until the next snapshot and then
runs the interpreter to the same program point.  Any side effects
of the IR code are redirected to the shadow heap and the shadow
stack.  Once the interpreter reaches the same program point we
compare the shadow stack & heap with the actual stack & heap.  This
will hopefully make it easier to find bugs in the future.

Other changes:

  * J->slots now keeps track of whether the value was written or simply
    named.  The heuristic of checking whether the value was constructed
    using SLOAD was a hack.

  * Add single-stepping mode for the interpreter (used by
    self-verification).

  * Add flags to enable/disable trace optimisations on a finer level.

  * Add different guards for return address and info table checks.
    Currently they are operationally equivalent to EQ, but that may
    change in the future.