Fix writing 32 bit immediates as 64 bit values.
On x64 there is no instruction for writing a 64 bit immediate directly
into memory.  We have to load it into a register first, this is what
we do most of the time on OS X since all addresses are at least 33
bits wide.

On Linux we can allocate below the 4G boundary and hence can use an
instruction that writes a 32 bit constant directly to memory.
However, the default operand width is 32 bits instead of 64 bits.
This patch fixes the encoding of this optimised instruction encoding
by making sure we actually write a 64 bit value to memory.