Store info tables in capability.
This is needed for future refactorings.  The `engine` function becomes
`engine_impl` and takes an operation mode.  The reason this is done
this way is because the interpreter uses direct threading and we can
only call these labels within this function.  So we need one mode,
where the engine operates as the interpreter as intended, and one mode
where it stores pointer to the dispatch tables in the capability.  No,
it's not pretty.