Switch to new hot trace detection style.
Each non-local branch (CALL[T], EVAL, RET) now calls into a function
that determines whether trace recording should start at the target.

This also changes the policy of where traces may start. Until now we
had a counter for each branch target and thus a trace could start
anywhere. This has been changed so that only targets of a backwards
branch can start a trace (as any loop must contain such a branch).

Note that this still creates two traces for non tail recursive
functions.  For example,

    foo:  FUNC
          ...
      A:  r = CALL foo(x, y)
          ...
      B:  RET1 z

This will create a trace <foo-to-A> because the branch from A to foo
is a backwards branch (with increasing stack), and a trace <A-to-B>
because RET1 will branch to A (with decreasing stack).