Fix recorder/interpreter sign mismatch.
The trace recorder currently executes before the interpreter processes
these instructions. This to find out which way a branch goes it has to
evaluate the condition itself. This fixes a mismatch between the two
implementations.

We probably need a nicer way of detecting which way a branch goes.
For EVAL/CALL we need the target's info table, though, and for RETs we
need the return address.