Fix the encoder bug exposed by Bc.Bc0014.
The problem was that we assumed that the default branch for a case was
always the fall-through target. This was never tested for and is
indeed quite hard to ensure in the current setup. This change
therefore always emits a branch to the default label which may be
redundant, but never wrong.  I.e., a the case expression

    [| case E1 of tag1 -> E2; _ -> E3 |]

gets translated and encoded as:

      t1 = [| E1 |]
      CASE t1
        tag1: ->Label2
      JMP ->Label3
    Label2:
      [| E2 |]
    Label3:
      [| E3 |]

If the order of E2 and E3 is swapped the JMP instruction behaves like
a no-op.