Fix Bc.Bc0014 again.  Properly this time.
The previous fix did not correctly calculate the size of the dispatch
table.  The dispatch table size is the maximum returned tag, not the
maximum tag we happen to match on.  To get the maximum tag we use the
type information from Core.  This requires that non-trivial pattern
matches are always on algebraic types.  So far that seems to be the
case for all our library code.

Note: It may make sense to introduce a specialised instruction ISTAG
that corresponds to the case:

    case E1 of { A -> E2; _ -> E3 }

This seems to be a rather common occurrence.