Initialise built-in closures in loader.
These closures are currently global variables to avoid unnecessary
memory access when using them.  In practise, though, they are more
members of the memory manager as they point into memory managed by the
memory allocator.  If at some point we want to support multiple VMs
running in the same process, this would need to be changed.