Change the way we store 64 bit literals.
Literals that require more than 32 bits cannot be stored inline in an
IR instruction. Previously, we used a separate array of 64 bit numbers
and stored an index into that array in the IR instruction. This was
rather messy to deal with, so we now emit two IR instructions
KWORD/KWORDHI which store the low and high parts, respectively. This
has the same memory overhead and both IR instructions are now always
adjacent in memory which improves locality.