Switch back to old CALL[T], *FUNC behaviour.
Before this change partial and over-applications were handled by
FUNC/IFUNC. This was rather messy and bug-prone. It also didn't work
well with the JIT. The main reason for these issues is that we needed
to transfer argument information (no. of arguments + pointer mask)
from caller to the FUNC. This was done using the "register" that would
normally hold the predecoded D parameter. While seemingly very clever
and low-overhead, it caused many issues elsewhere:

  - When CALL triggers some other code to be executed or even a mode
    switch, we have to be careful that the right D arrives at the
    FUNC instruction.

  - Compiled traces somehow have to deal with the D parameter, too.
    For non-exact calls the trace would have to turn the frame setup
    code into compiled code. That, however, specialises the trace to
    the individual call site.

  - We could handle the stack frame transformation before entering
    the trace, but that would mean things work differently in the
    interpreter and trace compiler. Usually not a good idea.

So, instead the caller takes care of handling all the
cases (PAP,THUNK,FUN{exact,over,partial}). This keeps the complexity
in one place, which overall makes it nicer to deal with.