Don't zero out blocks in performance mode.
When developing, it's useful to fill empty blocks with a dummy value
to detect when we're accessing stale memory. It does cost a fair bit
of performance, so we better turn this off when benchmarking. For
example, when running Bench.SumFromTo1 with N = 100000000 it reduced
GC from 500ms to 3ms! We're now actually faster that GHC on this
particular micro-benchmark.