Abort recording if we hit a heap block overflow.
Upon encountering a heap-overflow during an allocation instruction we
adjust the heap pointer and then try to execute the instruction again.
This caused problems when executing the instruction in recording mode.
Rather than handle this case in the recorder we simply abort recording
and retry another time (at least for now).