Add support for quick heap check failures.
Whenever we reach fill up an allocation buffer we get a heap check
failure.  So far this meant falling back to the interpreter, executing
one iteration in the interpreter and then re-enter the trace.  If a GC
is necessary this approach is fine.  Most of the time, however, we
just need to grab a new buffer and update the heap pointer and the
heap limit and nothing else.  Falling back to the interpreter for this
very frequent case causes a fair amount of overhead.

This patch uses a special handler that re-enters the trace if we only
needed to grab a new buffer and only exits the trace if GC is
required.

Benchmarks (elapsed MUT time):

               ghc-7.0.3 -O2 | lcvm (old) | lcvm (new)
              -----------------------------------------
  SumFromTo1       2.24s         2.42s       2.08s
  SumSquare1       2.50s         2.64s       2.08s

Yes, it's weird that lcvm now takes the same time, but they generate a
different number of traces, so I'm not accidentally running the wrong
code.