Store return results on returning function's stack
The function return result is no longer stored in a the current
thread, but instead on the stack itself.  This makes it
straight-forward to implement multiple return results (with some help
from the bytecode compiler it becomes a no-op -- same as tail call).
Tracing support is not yet implemented but this will also become
easier.