Fix stack corruption issue; tests in separate commit.
The parallel assignment code used a bad register as a temporary
register because it thought it was free.  The fix is that whenever we
emit an assignment "dest <- source" we mark "dest" as free, and mark
"source" as used.  Kind of obvious, but it only bit us under certain
circumstances.

There are a few other issues lurking in the parallel assign code, so
I'm adding a proper random-testing based stress test in a separate
commit.