Avoid overallocation when exiting to a side trace
Since we combine multiple heap checks into a single heap check, we
may end up allocating too much memory if execution leaves the trace
at a side exit.  For traces with poor completion rate this may cause
significant amounts of wasted memory and hence more frequent garbage
collections.  We thus keep track of this wastage and decrement the
heap pointer by the appropriate amount at the beginning of the side
trace.

Allocation rate of Spectral.Constrainst is now no longer changes if
we modify trace selection parameters. There is actually still a small
amount of variation, but that is most likely because we do not adjust
the heap pointer when switching to the interpreter (yet).