Get things to compile with full static optimisations
There was some code that forced -O0 to the GHC frontend due to earlier
issues with the bytecode compiler.  There are still some weird things
going on, but Spectral.Constraints is working now.

I had to create local definitions of `tail` and `!!` because there
were references to `tail1` and `!!1` which I could not recreate.
Presumably, this is some SpecConstr magic which somehow fired when
compiled using GHC, but not our frontend.  Maybe requires some
parameter tweaking when compiling our version.