Do not reuse spill slots.  It's subtly wrong.
Counter example in `docs/reuse_spills.markdown`.  The constraint is
that if a variable is spilled, the spill slot must be different from
any spill slots used throughout its lifetime.

Non-solution: Store the current register assignments in the snapshot.
Downside is that this does not allow weak references and thus may keep
values alive even if they are only used in side exits.