Handle mmap returning unaligned chunks
OS X 10.7+ do address space layout randomization (ASLR) and (mostly)
ignore the hint to mmap.  We now ask for a larger region and munmap
the stuff outside the aligned area.  I haven't yet tested how this
behaves when we try to get multiple adjacent blocks.  It's quite
possible that this will waste quite a bit of address space.  It's
probably not an issue given we have 48 bits of address space.