Add context type to Context
A BindC context can only come from a `case` expression where we can
always get the result type of the case alternatives from the case
binder.  We thus propagate the case type with the context to avoid
having to compute the result type, which seems to be quite
fragile (likely to break with a new version of GHC).

TODO: Verify that even if the binder is a wildcard its type
information is correct.